tools to read and write sav files to and from dict
